ex5/project/backend/requirements.txt
fastapi
uvicorn[standard]
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pandas
numpy
scipy
matplotlib
reportlab
jinja2
python-multipart
